+++ tests/test_wrapperutils.py
@@ -13,6 +13,7 @@ import            textwrap
 try: from unittest import mock
 except ImportError: import mock
 
+import six
 import pytest
 
 import numpy as np
@@ -25,7 +26,7 @@ import fsl.data.image            as fslimage
 import fsl.wrappers.wrapperutils as wutils
 
 
-from . import mockFSLDIR
+from . import mockFSLDIR, cleardir, checkdir
 from .test_run import mock_submit
 
 
@@ -309,25 +310,278 @@ def test_fileOrImage():
         assert np.all(result.get_data()[:] == expected)
 
 
+def test_fileOrThing_sequence():
+
+    @wutils.fileOrArray('arrs', 'out')
+    def func(arrs, out):
+
+        if isinstance(arrs, six.string_types):
+            arrs = [arrs]
+
+        arrs = [np.loadtxt(a) for a in arrs]
+        res  = np.sum(arrs, axis=0)
+
+        np.savetxt(out, res)
+
+    inputs  = [np.random.randint(1, 10, (3, 3)) for i in range(4)]
+    infiles = ['input{}.txt'.format(i) for i in range(len(inputs))]
+    exp     = np.sum(inputs, axis=0)
+
+    with tempdir.tempdir():
+
+        for ifile, idata in zip(infiles, inputs):
+            np.savetxt(ifile, idata)
+
+        func(inputs, 'result.txt')
+        assert np.all(np.loadtxt('result.txt') == exp)
+
+        assert np.all(func(inputs, wutils.LOAD)['out'] == exp)
+
+        func(inputs[0], 'result.txt')
+        assert np.all(np.loadtxt('result.txt') == inputs[0])
+
+        assert np.all(func(inputs[0], wutils.LOAD)['out'] == inputs[0])
+
+        func(infiles, 'result.txt')
+        assert np.all(np.loadtxt('result.txt') == exp)
+
+        assert np.all(func(infiles, wutils.LOAD)['out'] == exp)
+
+        func(infiles[0], 'result.txt')
+        assert np.all(np.loadtxt('result.txt') == inputs[0])
+
+        assert np.all(func(infiles[0], wutils.LOAD)['out'] == inputs[0])
+
+
+def test_fileOrThing_outprefix():
+
+    @wutils.fileOrImage('img', outprefix='output_base')
+    def basefunc(img, output_base):
+        img = nib.load(img).get_data()
+
+        out1 = nib.nifti1.Nifti1Image(img * 5,  np.eye(4))
+        out2 = nib.nifti1.Nifti1Image(img * 10, np.eye(4))
+
+        nib.save(out1, '{}_times5.nii.gz' .format(output_base))
+        nib.save(out2, '{}_times10.nii.gz'.format(output_base))
+
+
+    with tempdir.tempdir() as td:
+        img  = nib.nifti1.Nifti1Image(np.array([[1, 2], [3, 4]]), np.eye(4))
+        exp1 = img.get_data() * 5
+        exp2 = img.get_data() * 10
+        nib.save(img, 'img.nii')
+
+        basefunc('img.nii', 'myout')
+        assert np.all(nib.load('myout_times5.nii.gz') .get_data() == exp1)
+        assert np.all(nib.load('myout_times10.nii.gz').get_data() == exp2)
+        cleardir(td, 'myout*')
+
+        basefunc(img, 'myout')
+        assert np.all(nib.load('myout_times5.nii.gz') .get_data() == exp1)
+        assert np.all(nib.load('myout_times10.nii.gz').get_data() == exp2)
+        cleardir(td, 'myout*')
+
+        res = basefunc(img, 'myout', myout_times5=wutils.LOAD)
+        assert np.all(res['myout_times5'].get_data() == exp1)
+        cleardir(td, 'myout*')
+
+        res = basefunc(img, 'myout', myout_times10=wutils.LOAD)
+        assert np.all(res['myout_times10'].get_data() == exp2)
+        cleardir(td, 'myout*')
+
+        res = basefunc(img, 'myout', myout=wutils.LOAD)
+        assert np.all(res['myout_times5'] .get_data() == exp1)
+        assert np.all(res['myout_times10'].get_data() == exp2)
+        cleardir(td, 'myout*')
+
+
+def test_fileOrThing_outprefix_differentTypes():
+
+    @wutils.fileOrImage('img', outprefix='outpref')
+    def func(img, outpref):
+
+        img  = nib.load(img)
+        img  = nib.nifti1.Nifti1Image(img.get_data() * 2, np.eye(4))
+        text = '1234567890'
+
+        nib.save(img, '{}_image.nii.gz' .format(outpref))
+
+        with open('{}_text.txt'.format(outpref), 'wt') as f:
+            f.write(text)
+
+    with tempdir.tempdir() as td:
+        img  = nib.nifti1.Nifti1Image(np.array([[1, 2], [3, 4]]), np.eye(4))
+        expi = img.get_data() * 2
+        expt = '1234567890'
+
+        func(img, 'myout')
+        assert np.all(nib.load('myout_image.nii.gz') .get_data() == expi)
+        with open('myout_text.txt', 'rt') as f:
+            assert f.read().strip() == expt
+        cleardir(td, 'myout*')
+
+        res = func(img, 'myout', myout_image=wutils.LOAD)
+        assert list(res.keys()) == ['myout_image']
+        assert np.all(res['myout_image'].get_data() == expi)
+        cleardir(td, 'myout*')
+
+        res = func(img, 'myout', myout=wutils.LOAD)
+        assert list(res.keys()) == ['myout_image']
+        assert np.all(res['myout_image'].get_data() == expi)
+        cleardir(td, 'myout*')
+
+        res = func(img, 'myout', myout_text=wutils.LOAD)
+        assert list(res.keys()) == []
+        cleardir(td, 'myout*')
+
+
+def test_fileOrThing_outprefix_directory():
+
+    import logging
+    logging.basicConfig()
+    logging.getLogger('fsl.wrappers').setLevel(logging.DEBUG)
+
+    @wutils.fileOrImage('img', outprefix='outpref')
+    def func(img, outpref):
+        img  = nib.load(img)
+        img2 = nib.nifti1.Nifti1Image(img.get_data() * 2, np.eye(4))
+        img4 = nib.nifti1.Nifti1Image(img.get_data() * 4, np.eye(4))
+
+        outdir = op.abspath('{}_imgs'.format(outpref))
+
+        os.mkdir(outdir)
+
+        nib.save(img2, op.join(outdir, 'img2.nii.gz'))
+        nib.save(img4, op.join(outdir, 'img4.nii.gz'))
+
+    with tempdir.tempdir() as td:
+        img  = nib.nifti1.Nifti1Image(np.array([[1, 2], [3, 4]]), np.eye(4))
+        exp2 = img.get_data() * 2
+        exp4 = img.get_data() * 4
+
+        res = func(img, 'myout')
+        assert len(res) == 0
+        checkdir(td,
+                 op.join('myout_imgs', 'img2.nii.gz'),
+                 op.join('myout_imgs', 'img4.nii.gz'))
+        cleardir(td, 'myout*')
+
+        res = func(img, 'myout', myout_imgs=wutils.LOAD)
+        assert len(res) == 2
+        assert np.all(res['myout_imgs/img2'].get_data() == exp2)
+        assert np.all(res['myout_imgs/img4'].get_data() == exp4)
+
+        res = func(img, 'myout', **{'myout_imgs/img2' : wutils.LOAD})
+        assert len(res) == 1
+        assert np.all(res['myout_imgs/img2'].get_data() == exp2)
+
+        res = func(img, 'myout', **{'myout_imgs/img' : wutils.LOAD})
+        assert len(res) == 2
+        assert np.all(res['myout_imgs/img2'].get_data() == exp2)
+        assert np.all(res['myout_imgs/img4'].get_data() == exp4)
+
+        os.mkdir('foo')
+        res = func(img, 'foo/myout')
+        assert len(res) == 0
+        checkdir(td,
+                 op.join('foo', 'myout_imgs', 'img2.nii.gz'),
+                 op.join('foo', 'myout_imgs', 'img4.nii.gz'))
+        cleardir(td, 'foo')
+
+        os.mkdir('foo')
+        res = func(img, 'foo/myout', **{'foo/myout' : wutils.LOAD})
+        assert len(res) == 2
+        assert np.all(res['foo/myout_imgs/img2'].get_data() == exp2)
+        assert np.all(res['foo/myout_imgs/img4'].get_data() == exp4)
+
+
 def test_chained_fileOrImageAndArray():
-    @wutils.fileOrImage('image')
-    @wutils.fileOrArray('array')
-    def func(image, array):
-        nib.load(image)
-        np.loadtxt(array)
+    @wutils.fileOrImage('image', 'outimage')
+    @wutils.fileOrArray('array', 'outarray')
+    def func(image, array, outimage, outarray):
+        image = nib.load(image)
+        array = np.loadtxt(array)
+
+        outimg = nib.nifti1.Nifti1Image(image.get_data() * 2, np.eye(4))
+
+        np.savetxt(outarray, array * 2)
+        outimg.to_filename(outimage)
 
     image = nib.nifti1.Nifti1Image(np.array([[1,  2], [ 3,  4]]), np.eye(4))
     array = np.array([[5, 6, 7, 8]])
+
+    expimg = nib.nifti1.Nifti1Image(image.get_data() * 2, np.eye(4))
+    exparr = array * 2
 
     with tempdir.tempdir():
 
         nib.save(image, 'image.nii')
         np.savetxt('array.txt', array)
 
-        func('image.nii', 'array.txt')
-        func('image.nii',  array)
-        func( image,      'array.txt')
-        func( image,       array)
+        func('image.nii', 'array.txt', 'outimg.nii', 'outarr.txt')
+        assert np.all(nib.load('outimg.nii').get_data() == expimg.get_data())
+        assert np.all(np.loadtxt('outarr.txt') == exparr)
+
+        func('image.nii', array, 'outimg.nii', 'outarr.txt')
+        assert np.all(nib.load('outimg.nii').get_data() == expimg.get_data())
+        assert np.all(np.loadtxt('outarr.txt') == exparr)
+
+        func( image, 'array.txt', 'outimg.nii', 'outarr.txt')
+        assert np.all(nib.load('outimg.nii').get_data() == expimg.get_data())
+        assert np.all(np.loadtxt('outarr.txt') == exparr)
+
+        func( image, array, 'outimg.nii', 'outarr.txt')
+        assert np.all(nib.load('outimg.nii').get_data() == expimg.get_data())
+        assert np.all(np.loadtxt('outarr.txt') == exparr)
+
+        res = func(image, array, wutils.LOAD, 'outarr.txt')
+        assert np.all(res['outimage'].get_data() == expimg.get_data())
+        assert np.all(np.loadtxt('outarr.txt') == exparr)
+
+        res = func(image, array, 'outimg.nii', wutils.LOAD)
+        assert np.all(nib.load('outimg.nii').get_data() == expimg.get_data())
+        assert np.all(res['outarray'] == exparr)
+
+        res = func(image, array, wutils.LOAD, wutils.LOAD)
+        assert np.all(res['outimage'].get_data() == expimg.get_data())
+        assert np.all(res['outarray'] == exparr)
+
+
+def test_fileOrThing_chained_outprefix():
+
+    import logging
+    logging.basicConfig()
+    logging.getLogger('fsl.wrappers').setLevel(logging.DEBUG)
+
+    @wutils.fileOrImage('image', 'outimage', outprefix='out')
+    @wutils.fileOrArray('array', 'outarray', outprefix='out')
+    def func(image, array, out, **kwa):
+
+        image = nib.load(image)
+        array = np.loadtxt(array)
+
+        outimg = nib.nifti1.Nifti1Image(image.get_data() * 2, np.eye(4))
+        outarr = array * 2
+
+        np.savetxt('{}_array.txt'.format(out), outarr)
+        outimg.to_filename('{}_image.nii'.format(out))
+
+    image = nib.nifti1.Nifti1Image(np.array([[1,  2], [ 3,  4]]), np.eye(4))
+    array = np.array([[5, 6, 7, 8]])
+
+    expimg = nib.nifti1.Nifti1Image(image.get_data() * 2, np.eye(4))
+    exparr = array * 2
+
+    with tempdir.tempdir():
+
+        func(image, array, 'myout')
+        assert np.all(nib.load('myout_image.nii').get_data() == expimg.get_data())
+        assert np.all(np.loadtxt('myout_array.txt') == exparr)
+
+        res = func(image, array, wutils.LOAD)
+        assert np.all(res['out_image'].get_data() == expimg.get_data())
+        assert np.all(res['out_array'] == exparr)
 
 
 def test_cmdwrapper():